% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install packages from the natverse GitHub repository.}
\usage{
install(
  pkgname = "natverse",
  dependencies = TRUE,
  confirm.dependencies.upgrade = "never",
  ...
)
}
\arguments{
\item{pkgname}{Package name inside the `natverse` repository.}

\item{dependencies}{Which dependencies you want to install see
\code{\link[remotes]{install_github}}.}

\item{confirm.dependencies.upgrade}{Whether to ask to install dependencies of
natmanager. See the \code{upgrade} argument of
\code{\link[remotes]{install_github}} for details. The default will never
ask to upgrade dependencies but just go ahead and do this is as necessary.}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}}.}
}
\description{
This will check to see if the GITHUB_PAT is set and will provide instructions
to the user to set it. Further if the GITHUB_PAT is set, it will start
installing the dependencies of the `natverse` package.
}
\examples{
\dontrun{
natverse::install()

}
}
