% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_repo}
\alias{list_repo}
\title{List all the repos inside a particular GitHub organisation}
\usage{
list_repo(orgname = "natverse")
}
\arguments{
\item{orgname}{Name of the GitHub organization}
}
\value{
Character vector of repository names
}
\description{
by default this will list all the repositories inside the
  `natverse` organization.
}
\examples{
\donttest{
natverse::list_repo()
}
}
