% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{selfupdate}
\alias{selfupdate}
\title{Update the `natmanager` package itself.}
\usage{
selfupdate(
  source = c("GITHUB", "CRAN"),
  upgrade.dependencies = "always",
  force = FALSE,
  ...
)
}
\arguments{
\item{source}{Location from which to obtain a newer version of natmanager.
Defaults to GITHUB since this may well have a newer version than the CRAN
package repository.}

\item{upgrade.dependencies}{Whether to install dependencies of natmanager.
See the \code{upgrade} argument of \code{\link[remotes]{install_github}}
for details. The default will go ahead and always do this is as necessary.}

\item{force}{Force self update even if there doesn't seem to be an update
(default \code{FALSE})}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}}
or \code{\link[remotes]{install_cran}}.}
}
\value{
Logical indicating whether an update was required (invisibly).
}
\description{
Update the `natmanager` package itself.
}
\examples{
\dontrun{
natmanager::selfupdate()
}
}
\seealso{
\code{\link{install}}
}
