% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{This will update the `natmanager` package itself.}
\usage{
update(source = c("GITHUB", "CRAN"), ...)
}
\arguments{
\item{source}{Location to look for updating the package from, can be 'GITHUB' or 'CRAN'.}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}} or \code{\link[remotes]{install_cran}}.}
}
\description{
This will update the `natmanager` package itself.
}
\examples{
\dontrun{
library('natmanager')
update()
}
}
