% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install natverse packages from GitHub}
\usage{
install(
  collection = c("core", "natverse"),
  dependencies = TRUE,
  upgrade.dependencies = "always",
  ...
)
}
\arguments{
\item{collection}{The collection of natverse packages that you would like to
install. The current optiond are \code{core}, the default, or
\code{natverse}. See Description for more information.}

\item{dependencies}{Which dependencies you want to install see
\code{\link[remotes]{install_github}}.}

\item{upgrade.dependencies}{Whether to install dependencies of natmanager.
See the \code{upgrade} argument of \code{\link[remotes]{install_github}}
for details. The default will go ahead and always do this is as necessary.}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}}.}
}
\description{
There are currently two collections of nat packages

  \itemize{

  \item \code{core} a minimal install that can use to get started with nat
  and already solve many problems

  \item \code{natverse} a powerful "batteries included" distribution with all
  mature packages in the natverse.

  }

  Since the \code{natverse} option will install many packages from GitHub,
  you need to have a GitHub account and personal access token (GITHUB_PAT).
  Install will check to see if you have a \code{GITHUB_PAT} already and, if
  not, walk you through the steps of setting one up.
}
\examples{
\dontrun{
# install core packages to try out the core natverse
natmanager::install('core')

# Full "batteries included" installation with all packages
natmanager::install('natverse')
}
}
