% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install packages inside the natverse repository.}
\usage{
install(pkgname, ...)
}
\arguments{
\item{pkgname}{Package name inside the `natverse` repository.}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}}.}
}
\description{
This will check to see if the GITHUB_PAT is set and will provide instructions to the user to set it
Further, if the GITHUB_PAT is set it will start installing the dependencies of the `natverse` package.
}
\examples{
\dontrun{
install()
}
}
