% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_repo}
\alias{list_repo}
\title{List all the repos inside a particular ogranization inside the natverse repository.}
\usage{
list_repo(orgname = "natverse")
}
\arguments{
\item{orgname}{Name of the organization in github to list repositories from.}
}
\value{
List of repositories
}
\description{
This will list all the repositories inside the `natverse` organization.
}
\examples{
\dontrun{
library('natmanager')
list_repo()
}
}
