% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{selfupdate}
\alias{selfupdate}
\title{Update the `natmanager` package itself.}
\usage{
selfupdate(source = c("GITHUB", "CRAN"), dependencies.upgrade = "never", ...)
}
\arguments{
\item{source}{Location from which to obtain a newer version of natmanager.
Defaults to GITHUB since this may well have a newer version than the CRAN
packgae repository.}

\item{dependencies.upgrade}{Whether to ask to install dependencies of
natmanager. See the \code{upgrade} argument of
\code{\link[remotes]{install_github}} for details.}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}}
or \code{\link[remotes]{install_cran}}.}
}
\description{
Update the `natmanager` package itself.
}
\examples{
\dontrun{
natmanager::selfupdate()
}
}
\seealso{
\code{\link{install}}
}
