% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{selfupdate}
\alias{selfupdate}
\title{Update the `natmanager` package itself.}
\usage{
selfupdate(source = c("GITHUB", "CRAN"), upgrade.dependencies = "always", ...)
}
\arguments{
\item{source}{Location from which to obtain a newer version of natmanager.
Defaults to GITHUB since this may well have a newer version than the CRAN
packgae repository.}

\item{upgrade.dependencies}{Whether to install dependencies of natmanager.
See the \code{upgrade} argument of \code{\link[remotes]{install_github}}
for details. The default will go ahead and always do this is as necessary.}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}}
or \code{\link[remotes]{install_cran}}.}
}
\description{
Update the `natmanager` package itself.
}
\examples{
\dontrun{
natmanager::selfupdate()
}
}
\seealso{
\code{\link{install}}
}
