% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update the `natmanager` package itself.}
\usage{
update(source = c("GITHUB", "CRAN"), ...)
}
\arguments{
\item{source}{Location to check package updates. Defaults to GITHUB since
this may well have a newer version than the CRAN packgae repository.}

\item{...}{extra arguments to pass to \code{\link[remotes]{install_github}}
or \code{\link[remotes]{install_cran}}.}
}
\description{
Update the `natmanager` package itself.
}
\examples{
\dontrun{
library('natmanager')
update()
}
}
\seealso{
\code{\link{install}}
}
